code cleanup and minor refactor
